refactor: rename raft_log_2 to raft_log_v004; move struct to seperate files
