+++ src/meta/binaries/metactl/import.rs
@@ -28,8 +28,9 @@ use std::sync::Mutex;
 use databend_common_meta_raft_store::config::RaftConfig;
 use databend_common_meta_raft_store::key_spaces::RaftStoreEntry;
 use databend_common_meta_raft_store::key_spaces::SMEntry;
-use databend_common_meta_raft_store::log::raft_log_2;
 use databend_common_meta_raft_store::ondisk::DataVersion;
+use databend_common_meta_raft_store::raft_log_v004::log_store_meta;
+use databend_common_meta_raft_store::raft_log_v004::util;
 use databend_common_meta_raft_store::sm_v003::adapter::SnapshotUpgradeV002ToV003;
 use databend_common_meta_raft_store::sm_v003::write_entry::WriteEntry;
 use databend_common_meta_raft_store::sm_v003::SnapshotStoreV003;
@@ -375,10 +376,10 @@ async fn init_new_cluster(
     // Reset node id
     {
         let mut log = sto.log.write().await;
-        log.save_user_data(Some(raft_log_2::LogStoreMeta {
+        log.save_user_data(Some(log_store_meta::LogStoreMeta {
             node_id: Some(args.id),
         }));
-        raft_log_2::blocking_flush(&mut log).await?;
+        util::blocking_flush(&mut log).await?;
     }
 
     Ok(())
